move files to work on both app and page routers
