+++ next-app/app/_components/layout/Footer.tsx
