+++ next-app/pages/prev_report.tsx
@@ -1,7 +1,7 @@
 import { useState, ChangeEvent, FormEvent } from 'react'
 import Layout from '@/components/Layout/Layout'
 import Container from '@/components/Container'
-import useSendingStatus from '@/hooks/useSendingStatus'
+import useSendingStatus from '@/app/_hooks/useSendingStatus'
 import SendButton from '@/components/SendButton'
 
 export default function Report() {
